java.lang.ClassCastException: java.lang.Float cannot be cast to java.lang.Double (talkback group 362)
git-svn-id: http://svn.code.sf.net/p/jebl/code/trunk@882 6721810a-1124-0410-938e-bb277542c816

+++ jebl/src/jebl/gui/trees/treeviewer/painters/BasicLabelPainter.java
@@ -401,7 +401,7 @@ public class BasicLabelPainter extends AbstractPainter<Node> {
             final String fontSizePrefKey = getTitle() + "_fontsize";
             final float fontsize = PREFS.getFloat(fontSizePrefKey, taxonLabelFont.getSize());
             setFontSize(fontsize, false);
-            fontSizeSpinner.setValue(fontsize);
+            fontSizeSpinner.setValue((double)fontsize);
